- Fixes for pull request
