+++ bootstrap/build.gradle
@@ -23,7 +23,7 @@ dependencies {
     final def springBootVersion = '2.1.3.RELEASE'
     final def kotlinVersion = '1.3.10'
 
-    /*implementation project(":notary-commons")*/
+    implementation project(":notary-commons")
 
     // kotlin related
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlinVersion"
@@ -60,7 +60,7 @@ dependencies {
     compile 'com.google.protobuf:protobuf-java-util:3.6.1'
     compile 'org.bitcoinj:bitcoinj-core:0.14.7'
     compile 'org.web3j:crypto:4.0.4'
-    
+
     testCompile "org.springframework.boot:spring-boot-starter-test:$springBootVersion"
 
 }
