+++ bootstrap/src/main/kotlin/jp/co/soramitsu/bootstrap/config/BootstrapConfig.kt
@@ -1,0 +1,18 @@
+package jp.co.soramitsu.bootstrap.config
+
+import jp.co.soramitsu.bootstrap.genesis.d3.D3TestGenesisFactory
+import jp.co.soramitsu.bootstrap.genesis.GenesisInterface
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import springfox.documentation.swagger2.annotations.EnableSwagger2
+
+@Configuration
+@EnableSwagger2
+class BootstrapConfig {
+
+    @Bean
+    fun genesisFactories(): List<jp.co.soramitsu.bootstrap.genesis.GenesisInterface> {
+        return listOf(D3TestGenesisFactory())
+    }
+}
+
