+++ bootstrap/src/main/kotlin/jp/co/soramitsu/bootstrap/dto/block/GenesisBlock.kt
@@ -1,0 +1,32 @@
+package jp.co.soramitsu.bootstrap.dto.block
+
+import java.util.HashMap
+import com.fasterxml.jackson.annotation.JsonAnyGetter
+import com.fasterxml.jackson.annotation.JsonAnySetter
+import com.fasterxml.jackson.annotation.JsonIgnore
+import com.fasterxml.jackson.annotation.JsonInclude
+import com.fasterxml.jackson.annotation.JsonProperty
+import com.fasterxml.jackson.annotation.JsonPropertyOrder
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonPropertyOrder("blockV1")
+class GenesisBlock {
+
+    @JsonProperty("blockV1")
+    @get:JsonProperty("blockV1")
+    @set:JsonProperty("blockV1")
+    var blockV1: jp.co.soramitsu.bootstrap.dto.block.BlockV1? = null
+    @JsonIgnore
+    private val additionalProperties = HashMap<String, Any>()
+
+    @JsonAnyGetter
+    fun getAdditionalProperties(): Map<String, Any> {
+        return this.additionalProperties
+    }
+
+    @JsonAnySetter
+    fun setAdditionalProperty(name: String, value: Any) {
+        this.additionalProperties[name] = value
+    }
+
+}
