+++ bootstrap/src/main/kotlin/jp/co/soramitsu/bootstrap/error/Errors.kt
@@ -1,0 +1,6 @@
+package jp.co.soramitsu.bootstrap.error
+
+import java.lang.RuntimeException
+
+class AccountException(message:String) : RuntimeException(message)
+class IrohaPublicKeyError(message:String) : RuntimeException(message)
