--- bootstrap/src/main/kotlin/jp/co/soramitsu/notary/bootstrap/dto/block/CreateDomain.kt
@@ -1,36 +1,0 @@
-package jp.co.soramitsu.notary.bootstrap.dto.block
-
-import java.util.HashMap
-import com.fasterxml.jackson.annotation.JsonAnyGetter
-import com.fasterxml.jackson.annotation.JsonAnySetter
-import com.fasterxml.jackson.annotation.JsonIgnore
-import com.fasterxml.jackson.annotation.JsonInclude
-import com.fasterxml.jackson.annotation.JsonProperty
-import com.fasterxml.jackson.annotation.JsonPropertyOrder
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder("domainId", "defaultRole")
-class CreateDomain {
-
-    @JsonProperty("domainId")
-    @get:JsonProperty("domainId")
-    @set:JsonProperty("domainId")
-    var domainId: String? = null
-    @JsonProperty("defaultRole")
-    @get:JsonProperty("defaultRole")
-    @set:JsonProperty("defaultRole")
-    var defaultRole: String? = null
-    @JsonIgnore
-    private val additionalProperties = HashMap<String, Any>()
-
-    @JsonAnyGetter
-    fun getAdditionalProperties(): Map<String, Any> {
-        return this.additionalProperties
-    }
-
-    @JsonAnySetter
-    fun setAdditionalProperty(name: String, value: Any) {
-        this.additionalProperties[name] = value
-    }
-
-}
