--- bootstrap/src/main/kotlin/jp/co/soramitsu/notary/bootstrap/dto/block/CreateRole.kt
@@ -1,36 +1,0 @@
-package jp.co.soramitsu.notary.bootstrap.dto.block
-
-import java.util.HashMap
-import com.fasterxml.jackson.annotation.JsonAnyGetter
-import com.fasterxml.jackson.annotation.JsonAnySetter
-import com.fasterxml.jackson.annotation.JsonIgnore
-import com.fasterxml.jackson.annotation.JsonInclude
-import com.fasterxml.jackson.annotation.JsonProperty
-import com.fasterxml.jackson.annotation.JsonPropertyOrder
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder("roleName", "permissions")
-class CreateRole {
-
-    @JsonProperty("roleName")
-    @get:JsonProperty("roleName")
-    @set:JsonProperty("roleName")
-    var roleName: String? = null
-    @JsonProperty("permissions")
-    @get:JsonProperty("permissions")
-    @set:JsonProperty("permissions")
-    var permissions: List<String>? = null
-    @JsonIgnore
-    private val additionalProperties = HashMap<String, Any>()
-
-    @JsonAnyGetter
-    fun getAdditionalProperties(): Map<String, Any> {
-        return this.additionalProperties
-    }
-
-    @JsonAnySetter
-    fun setAdditionalProperty(name: String, value: Any) {
-        this.additionalProperties[name] = value
-    }
-
-}
