--- bootstrap/src/main/kotlin/jp/co/soramitsu/notary/bootstrap/dto/block/Payload_.kt
@@ -1,32 +1,0 @@
-package jp.co.soramitsu.notary.bootstrap.dto.block
-
-import java.util.HashMap
-import com.fasterxml.jackson.annotation.JsonAnyGetter
-import com.fasterxml.jackson.annotation.JsonAnySetter
-import com.fasterxml.jackson.annotation.JsonIgnore
-import com.fasterxml.jackson.annotation.JsonInclude
-import com.fasterxml.jackson.annotation.JsonProperty
-import com.fasterxml.jackson.annotation.JsonPropertyOrder
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder("reducedPayload")
-class Payload_ {
-
-    @JsonProperty("reducedPayload")
-    @get:JsonProperty("reducedPayload")
-    @set:JsonProperty("reducedPayload")
-    var reducedPayload: ReducedPayload? = null
-    @JsonIgnore
-    private val additionalProperties = HashMap<String, Any>()
-
-    @JsonAnyGetter
-    fun getAdditionalProperties(): Map<String, Any> {
-        return this.additionalProperties
-    }
-
-    @JsonAnySetter
-    fun setAdditionalProperty(name: String, value: Any) {
-        this.additionalProperties[name] = value
-    }
-
-}
