--- bootstrap/src/main/kotlin/jp/co/soramitsu/notary/bootstrap/dto/block/Peer.kt
@@ -1,36 +1,0 @@
-package jp.co.soramitsu.notary.bootstrap.dto.block
-
-import java.util.HashMap
-import com.fasterxml.jackson.annotation.JsonAnyGetter
-import com.fasterxml.jackson.annotation.JsonAnySetter
-import com.fasterxml.jackson.annotation.JsonIgnore
-import com.fasterxml.jackson.annotation.JsonInclude
-import com.fasterxml.jackson.annotation.JsonProperty
-import com.fasterxml.jackson.annotation.JsonPropertyOrder
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder("address", "peerKey")
-class Peer {
-
-    @JsonProperty("address")
-    @get:JsonProperty("address")
-    @set:JsonProperty("address")
-    var address: String? = null
-    @JsonProperty("peerKey")
-    @get:JsonProperty("peerKey")
-    @set:JsonProperty("peerKey")
-    var peerKey: String? = null
-    @JsonIgnore
-    private val additionalProperties = HashMap<String, Any>()
-
-    @JsonAnyGetter
-    fun getAdditionalProperties(): Map<String, Any> {
-        return this.additionalProperties
-    }
-
-    @JsonAnySetter
-    fun setAdditionalProperty(name: String, value: Any) {
-        this.additionalProperties[name] = value
-    }
-
-}
