--- bootstrap/src/main/kotlin/jp/co/soramitsu/notary/bootstrap/error/Errors.kt
@@ -1,6 +1,0 @@
-package jp.co.soramitsu.notary.bootstrap.error
-
-import java.lang.RuntimeException
-
-class AccountException(message:String) : RuntimeException(message)
-class IrohaPublicKeyError(message:String) : RuntimeException(message)
